pyproject.toml
[tool.poetry]
name = "pyhydrophone"
version = "0.1.6"
description = "python package to read and calibrate hydrophone data"
authors = ["Clea Parcerisas <clea.parcerisas@vliz.be>"]
license = "OSI Approved :: MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <3.12.0"
pandas = "^2.0.2"
requests = "^2.31.0"
scipy = "1.10.0"
soundfile = "^0.12.1"

# Testing tools
pyyaml = {version = "^6.0", optional = true}
pytest = {version = "^7.3.1", optional = true}
coverage = {version = "^7.2.5", optional = true}
python-dotenv = {version = "^1.0.0", optional = true}
pytest-cov = {version = "^4.1.0", optional = true}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "6.1.2"
numpydoc = "^1.5.0"
sphinx-rtd-theme = "^1.2.2"
sphinx-gallery = "^0.13.0"
pillow = "^10.0.0"
matplotlib = "^3.7.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests",]
pythonpath = ["."]